Final 1

+++ PU_CampusTour/Controllers/LocationController.cs
@@ -3,7 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
+//using System.Web.HttpRequestBase;
 using PU_CampusTour.Models;
+
+
 
 namespace PU_CampusTour.Controllers
 {
@@ -16,76 +19,71 @@ namespace PU_CampusTour.Controllers
         {
             return View();
         }*/
-
         
        public ActionResult location()
         {
-            
-            
+
+            if (Request.Cookies["Login"].Value.ToLower() == "true")
+            {
                 ViewBag.Message = "Done";
                 using (dbf522ec9140464818abf6a1c4013479aeEntities1 ctx = new dbf522ec9140464818abf6a1c4013479aeEntities1())
                 {
-                    // saving dummy values in database
-                   /* Place p = new Place()
-                    {
-                    //p.Place_Id = 1;
-                    Name = "Department of Biotechnology",
-                    Description = "The Department of Biotechnology started in 1962",
-                    Longitude = "16.90",
-                    Latitude = "1.078",
-                    Image = "E:/images/DE.jpg",
-                    //Category_Id = 1
-                    
-                    };
-
-                    ctx.AddToPlaces(p);
-                    ctx.SaveChanges();
-                    */
 
 
-                  // returning data in place class to be viewed
-                    return View(ctx.Places.ToList());  
+                    // returning data in place class to be viewed
+                    return View(ctx.Places.ToList());
                 }
+            }
+            else
+            {
+                return RedirectToAction("login","Login");
+            }
 
                    
         }
 
-       public ActionResult Request()
+       public ActionResult Requestplace()
        {
-
-
-           ViewBag.Message = "Done";
-           using (dbf522ec9140464818abf6a1c4013479aeEntities1 ctx = new dbf522ec9140464818abf6a1c4013479aeEntities1())
+           if (Request.Cookies["Login"].Value.ToLower() == "true")
            {
-               // saving dummy values in database
-               /* Place p = new Place()
-                {
-                //p.Place_Id = 1;
-                Name = "Department of Biotechnology",
-                Description = "The Department of Biotechnology started in 1962",
-                Longitude = "16.90",
-                Latitude = "1.078",
-                Image = "E:/images/DE.jpg",
-                //Category_Id = 1
-                    
-                };
 
-                ctx.AddToPlaces(p);
-                ctx.SaveChanges();
-                */
+               ViewBag.Message = "Done";
+               using (dbf522ec9140464818abf6a1c4013479aeEntities1 ctx = new dbf522ec9140464818abf6a1c4013479aeEntities1())
+               {
 
 
-               // returning data in place class to be viewed
-               return View(ctx.Places.ToList());
+
+                   // returning data in place class to be viewed
+                   return View(ctx.Places.ToList());
+               }
+           }
+           else
+           {
+               return RedirectToAction("login","Login");
            }
        }
 
         //
         // GET: /Add/
 
-        public ActionResult Add()
+        public ActionResult Add(int id)
         {
-            return View();
+            if (Request.Cookies["Login"].Value.ToLower() == "true")
+            {
+                if (id == 0)
+                {
+                    return View();
+                }
+                else
+                {
+                    ViewBag.id = id;
+                    return View();
+                }
+            }
+            else
+            {
+                return RedirectToAction("login","Login");
+            }
         }
 
 
@@ -102,35 +100,50 @@ namespace PU_CampusTour.Controllers
             int category = 0;
 
             // url = Request.Url[0];
-            name = c["Name"];
-            detail = c["Detail"];
-            longi = c["Longitude"];
-            lat = c["Latitude"];
-            category = Convert.ToInt32(c["Category"]);
+            
 
-
-            using (dbf522ec9140464818abf6a1c4013479aeEntities1 ctx = new dbf522ec9140464818abf6a1c4013479aeEntities1())
+            if (c["Name"] != "" && c["Detail"] != "" && c["Longitude"] != "" && c["Latitude"] != "" && c["Category"] != "")
             {
-                Place p = new Place()
+                name = c["Name"];
+                detail = c["Detail"];
+                longi = c["Longitude"];
+                lat = c["Latitude"];
+                category = Convert.ToInt32(c["Category"]);
+
+                using (dbf522ec9140464818abf6a1c4013479aeEntities1 ctx = new dbf522ec9140464818abf6a1c4013479aeEntities1())
                 {
-                    //p.Place_Id = 1;
-                    Name = name,
-                    Description = detail,
-                    Longitude = longi,
-                    Latitude = lat,
-                    Image = image,
-                    isApproved = "True"
+                    Place p = new Place()
+                    {
+                        //p.Place_Id = 1;
+                        Name = name,
+                        Description = detail,
+                        Longitude = longi,
+                        Latitude = lat,
+                        Image = image,
+                        isApproved = "True",
+                        Category_Id = category
 
-                    //Category_Id = 1
+                        //Category_Id = 1
 
-                };
+                    };
 
-                ctx.Places.Add(p);
-                ctx.SaveChanges();
+                    ctx.Places.Add(p);
+                    ctx.SaveChanges();
 
 
+                }
+                return RedirectToAction("Location", "Login");
             }
-            return RedirectToAction("Location", "Login");
+            else
+            {
+                ViewBag.MSG = "*Required fields are empty";
+                ViewBag.a1 = c["Name"];
+                ViewBag.a2 = c["Detail"];
+                ViewBag.a3 = c["Longitude"];
+                ViewBag.a4 = c["Latitude"];
+                ViewBag.a5= c["Category"];
+                return View();
+            }
 
 
         }
@@ -142,16 +155,24 @@ namespace PU_CampusTour.Controllers
         public ActionResult Edit(int id)
         {
             dbf522ec9140464818abf6a1c4013479aeEntities1 db = new dbf522ec9140464818abf6a1c4013479aeEntities1();
-            var q = from s in db.Places
-                    where s.Id == id
-                    select s;
-            //Pro_Class product = new Pro_Class();
-            
-            foreach (var t in q)
+
+            if (Request.Cookies["Login"].Value.ToLower() == "true")
             {
-                ViewBag.id = t.Id;
+                var q = from s in db.Places
+                        where s.Id == id
+                        select s;
+                //Pro_Class product = new Pro_Class();
+
+                foreach (var t in q)
+                {
+                    ViewBag.id = t.Id;
+                }
+                return View();
             }
-            return View();
+            else
+            {
+                return RedirectToAction("login", "Login");
+            }
         }
 
         [HttpPost]
@@ -166,42 +187,53 @@ namespace PU_CampusTour.Controllers
             int category = 0;
 
             // url = Request.Url[0];
-            name = c["Name"];
-            detail = c["Detail"];
-            longi = c["Longitude"];
-            lat = c["Latitude"];
-            category = Convert.ToInt32(c["Category"]);
 
-            int id = Convert.ToInt32(c["id"]);
-            using (dbf522ec9140464818abf6a1c4013479aeEntities1 ctx = new dbf522ec9140464818abf6a1c4013479aeEntities1())
+
+            if (c["Name"] != "" && c["Detail"] != "" && c["Longitude"] != "" && c["Latitude"] != "" && c["Category"] != "" && c["id"] != "")
             {
+                name = c["Name"];
+                detail = c["Detail"];
+                longi = c["Longitude"];
+                lat = c["Latitude"];
+                category = Convert.ToInt32(c["Category"]);
 
-                var q = from s in ctx.Places
-                        where s.Id == id
-                        select s;
-                foreach (var t in q)
+                int id = Convert.ToInt32(c["id"]);
+
+                using (dbf522ec9140464818abf6a1c4013479aeEntities1 ctx = new dbf522ec9140464818abf6a1c4013479aeEntities1())
                 {
-                    t.Name = name;
-                   
-                    t.Description = detail;
-                    
-                    t.Longitude = longi;
-                   
-                    t.Latitude = lat;
-                   
-                    t.Image = image;
 
-                    t.Category_Id = category;
-                    
-                    t.isApproved = "True";
+                    var q = from s in ctx.Places
+                            where s.Id == id
+                            select s;
+                    foreach (var t in q)
+                    {
+                        t.Name = name;
+
+                        t.Description = detail;
+
+                        t.Longitude = longi;
+
+                        t.Latitude = lat;
+
+                        t.Image = image;
+
+                        t.Category_Id = category;
+
+                        t.isApproved = "True";
+
+
+                    }
+                    ctx.SaveChanges();
 
 
                 }
-                ctx.SaveChanges();
-
-
+                return RedirectToAction("location", "Login");
             }
-            return RedirectToAction("location","Login");
+            else
+            {
+                ViewBag.MSG = "*Required fields are empty";
+                return View();
+            }
 
 
         }
@@ -209,16 +241,23 @@ namespace PU_CampusTour.Controllers
         public ActionResult Delete(int id)
         {
             dbf522ec9140464818abf6a1c4013479aeEntities1 db = new dbf522ec9140464818abf6a1c4013479aeEntities1();
-            var q = from s in db.Places
-                    where s.Id == id
-                    select s;
-            foreach (var t in q)
+            if (Request.Cookies["Login"].Value.ToLower() == "true")
             {
-                db.Places.Remove(t);
-            }
-            db.SaveChanges();
+                var q = from s in db.Places
+                        where s.Id == id
+                        select s;
+                foreach (var t in q)
+                {
+                    db.Places.Remove(t);
+                }
+                db.SaveChanges();
 
-            return RedirectToAction("location", "Login");
+                return RedirectToAction("location", "Login");
+            }
+            else
+            {
+                return RedirectToAction("login", "Login");
+            }
         }
 
        
@@ -227,23 +266,116 @@ namespace PU_CampusTour.Controllers
         public ActionResult DeleteAll(int[] deleteValues)
         {
             dbf522ec9140464818abf6a1c4013479aeEntities1 db = new dbf522ec9140464818abf6a1c4013479aeEntities1();
-            
-            int[] MyCheckboxes = deleteValues;
 
-            
-            foreach (var item in MyCheckboxes)
+            if (Request.Cookies["Login"].Value.ToLower() == "true")
             {
-                var q = from s in db.Places
-                        where s.Id == Convert.ToInt32(item)
-                        select s;
-                foreach (var t in q)
+                int[] MyCheckboxes = deleteValues;
+
+
+                foreach (var item in MyCheckboxes)
                 {
-                    db.Places.Remove(t);
+                    var q = from s in db.Places
+                            where s.Id == Convert.ToInt32(item)
+                            select s;
+                    foreach (var t in q)
+                    {
+                        db.Places.Remove(t);
+                    }
+                    db.SaveChanges();
                 }
-                db.SaveChanges();
+
+                return RedirectToAction("location", "Login");
+            }
+            else
+            {
+                return RedirectToAction("login", "Login");
+            }
+        }
+
+
+
+        public ActionResult Change_Password()
+        {
+            if (Request.Cookies["Login"].Value.ToLower() == "true")
+            {
+                
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("login", "Login");
+            }
+        }
+
+
+        [HttpPost]
+        public ActionResult Change_Password(FormCollection c)
+        {
+            string currpass = c["CP"];
+            string newpass=c["NP"];
+            string conpass = c["ConP"];
+            
+
+
+          if(currpass!="" && newpass!="" && conpass!="")
+          {
+            if (newpass == conpass)
+            {
+                using (dbf522ec9140464818abf6a1c4013479aeEntities1 db1 = new dbf522ec9140464818abf6a1c4013479aeEntities1())
+                {
+                    var q1 = from s1 in db1.Admins
+                             where s1.Password == currpass
+                             select s1;
+
+                    if (q1.Count()!=0)
+                    {
+                        foreach (var t1 in q1)
+                        {
+                            t1.Password = newpass;
+
+
+                        }
+                        db1.SaveChanges();
+
+
+                        ViewBag.msg = "Password has been updated successfully";
+
+                        return View();
+                    }
+                    else
+                    {
+                        ViewBag.msg = "*Incorrect current password";
+                        ViewBag.n1 = currpass;
+                        ViewBag.n2 = newpass;
+                        ViewBag.n3 = conpass;
+                        return View();
+                    }
+                }
+            
+            }
+            else
+            {
+                ViewBag.msg = "*New and Confirm Password doesnot match";
+                ViewBag.n1 = currpass;
+                ViewBag.n2 = newpass;
+                ViewBag.n3 = conpass;
+                return View();
+
             }
 
-            return RedirectToAction("location", "Login");
+              
+          }
+           else
+          {
+              ViewBag.msg = "*Required fields are empty";
+              ViewBag.n1 = currpass;
+              ViewBag.n2 = newpass;
+              ViewBag.n3 = conpass;
+              return View();
+          }
+                
+
+            
         }
     }
 }
